add def_print_tablebook
